tests/wpt/web-platform-tests/bluetooth/requestDevice/le-not-supported.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script>
'use strict';
const test_desc = 'Reject with NotFoundError if Bluetooth is not supported.';
const expected =
    new DOMException('Bluetooth Low Energy not available.', 'NotFoundError');

bluetooth_test(
    () => navigator.bluetooth.test.setLESupported(false).then(
        () => assert_promise_rejects_with_message(
            requestDeviceWithTrustedClick({acceptAllDevices: true}), expected,
            'Bluetooth Low Energy is not supported.')),
    test_desc);
</script>
